Added matching algorithms (Hopcraft-Karp, Kuhn-Munkres) to TeX document but without documentation yet. Moved matching sources to graph/matching from graph/maxflow Fixed spelling errors.
bk: 3e5bff5fV65L9wstcX4KfVsFL7lrDg

+++ numerical/prime_sieve.cpp
@@ -13,7 +13,7 @@ struct prime_sieve {
   char *v;
 
   prime_sieve( int n ) {
-    int sn = (int)sqrt(n);
+    int sn = (int)sqrt((double)n);
     n /= 2;
     v = (char*)malloc(n);
     memset( v, 1, n );
